common tracker & changed parameters
