+++ screen_tracking/tracker/hough_heuristics/tracking.py
@@ -3,6 +3,7 @@ import logging
 import cv2
 import numpy as np
 
+from screen_tracking.tracker.common_tracker import Tracker
 from screen_tracking.tracker.hough_heuristics.frontiers import (
     show_best,
     PreviousPoseFrontier,
@@ -20,14 +21,7 @@ from screen_tracking.tracker.hough_heuristics.utils import (
     screen_points_to_lines)
 
 
-class Tracker:
-    tracker_params = TrackerParams()
-
-    FRAMES_NUMBER_TO_TRACK = np.inf
-    INITIAL_FRAME = 1
-    PREVIOUS_GROUND_TRUTH = False
-
-    SHOW_EACH_SIDE = int(1e9)
+class HoughTracker(Tracker):
 
     def line_frontiers(self):
         last_lines = screen_points_to_lines(self.state.last_points)
@@ -113,92 +107,7 @@ class Tracker:
         intersections = screen_lines_to_points(lines)
         return intersections
 
-    def track(self):
-        self.state.productivity = {}
-        tracking_result = {}
-        initial_frame_number = self.INITIAL_FRAME
-        cap = cv2.VideoCapture(self.video_source)
-        frame_number = initial_frame_number
-        last_frame = 0
-        for i in range(frame_number):
-            ret, last_frame = cap.read()
-        last_frame = [last_frame]
-        last_points = [self.frame_pixels[frame_number]]
-        self.write_camera(tracking_result, last_points[0], frame_number)
-        predict_matrix = tracking_result[frame_number]
-
-        while cap.isOpened():
-            try:
-                if frame_number == initial_frame_number + self.FRAMES_NUMBER_TO_TRACK:
-                    break
-                frame_number += 1
-                print('\nFrame number: ', frame_number)
-                self.state.frame_number = frame_number
-                ret, frame = cap.read()
-                if not ret:
-                    break
-                points = self.get_points(frame, last_frame[0], last_points[0], predict_matrix)
-                if points is not None:
-                    self.write_camera(tracking_result, points, frame_number)
-                    predict_matrix = tracking_result[frame_number]
-
-                if self.PREVIOUS_GROUND_TRUTH:
-                    _, rotation, translation = cv2.solvePnP(
-                        self.model_vertices,
-                        self.frame_pixels[frame_number],
-                        self.camera_params,
-                        np.array([]),
-                        flags=self.tracker_params.PNP_FLAG
-                    )
-                    R_rodrigues = cv2.Rodrigues(rotation)[0]
-                    external_matrix = np.hstack((R_rodrigues, translation))
-                    predict_matrix = external_matrix
-
-                predicted_points = get_screen_points(self, predict_matrix)
-                last_points[0] = predicted_points
-                last_frame[0] = frame
-            except Exception as error:
-                logging.error('Tracker broken')
-                logging.exception(error)
-                break
-
-        print()
-        for key, value in self.state.productivity.items():
-            print(key, value[1])
-
-        return tracking_result
-
-    def show_list_best(self, side_frontiers):
-        frame = self.state.cur_frame.copy()
-        for i in range(4):
-            show_best(side_frontiers[i], frame=frame, no_show=i < 3, max_count=self.SHOW_EACH_SIDE)
-
-    def write_camera(self, tracking_result, pixels, frame):
-        _, rotation, translation = cv2.solvePnP(
-            self.model_vertices,
-            pixels,
-            self.camera_params,
-            np.array([]),
-            flags=self.tracker_params.PNP_FLAG
-        )
-        R_rodrigues = cv2.Rodrigues(rotation)[0]
-        external_matrix = np.hstack((R_rodrigues, translation))
-        tracking_result[frame] = external_matrix
+    tracker_params = TrackerParams()
 
     def __init__(self, model_vertices, camera_params, video_source, frame_pixels):
-        self.model_vertices = model_vertices
-        self.camera_params = camera_params
-        self.video_source = video_source
-        self.frame_pixels = frame_pixels
-        self.state = TrackerState()
-
-
-def track(model_vertices, camera_params, video_source, frame_pixels, write_callback, result_file, tracker_params=None):
-    tracker = Tracker(model_vertices, camera_params, video_source, frame_pixels)
-
-    if tracker_params is not None:
-        tracker.tracker_params = tracker_params
-
-    result = tracker.track()
-    write_callback(result)
-    return result
+        super().__init__(model_vertices, camera_params, video_source, frame_pixels)
